services/api/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
psycopg[binary]==3.2.3
qdrant-client==1.11.1
sentence-transformers==3.0.1
torch==2.3.1
pdfplumber==0.11.4
PyMuPDF==1.24.9          # keep this; remove 'pymupdf'
python-multipart==0.0.9
tenacity==8.5.0
httpx==0.27.2
readability-lxml==0.8.1  # works with lxml<5
lxml==4.9.4              # pin ONE lxml < 5 version
sse-starlette==2.1.0
